repoService/requirements.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.3.1
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
idna==2.10
mongoengine==0.20.0
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
xlwt==1.3.0